+++ BookStore.Api/Features/Models/BookResponseModel.cs
@@ -8,7 +8,7 @@ public class BookResponseModel
     public string ISBN { get; set; }
     public decimal Price { get; set; }
     public int Stock { get; set; }
-    public int GenreId { get; set; }
+    public long GenreId { get; set; }
     public int PageCount { get; set; }
     public DateTime PublishedDate { get; set; }
     public string Publisher { get; set; }
